closer to final steps of core implementation.
Changed most of the codebase to Pascal Case, some places might still have either camelCase or other arbitrary cases so I will have to find them and deal with them.

some of the tests are made to work with a test program I made in python, this is temporary I will make it more general in the future

Most of worker.go is done but we need to improve the content extraction to use goquery.